fix: adjust ledger swagger docs :bug:
