+++ components/ledger/internal/bootstrap/http/product.go
@@ -21,16 +21,17 @@ type ProductHandler struct {
 
 // CreateProduct is a method that creates product information.
 //
-//		@Summary        Create a Product
-//		@Description    Create a Product with the input payload
-//		@Tags           Products
-//		@Accept         json
-//		@Produce        json
-//	 @Param 			organization_id path string true "Organization ID"
-//	 @Param 			ledger_id path string true "Ledger ID"
-//		@Param          product body mmodel.CreateProductInput true "Product"
-//		@Success        200 {object} mmodel.Product
-//		@Router         /v1/organizations/{organization_id}/ledgers/{ledger_id}/products [post]
+//	@Summary		Create a Product
+//	@Description	Create a Product with the input payload
+//	@Tags			Products
+//	@Accept			json
+//	@Produce		json
+//	@Param			organization_id	path		string						true	"Organization ID"
+//	@Param			ledger_id		path		string						true	"Ledger ID"
+//	@Param			product			body		mmodel.CreateProductInput	true	"Product"
+//	@Param			Midaz-Id		header		string							false	"Request ID"
+//	@Success		200				{object}	mmodel.Product
+//	@Router			/v1/organizations/{organization_id}/ledgers/{ledger_id}/products [post]
 func (handler *ProductHandler) CreateProduct(i any, c *fiber.Ctx) error {
 	ctx := c.UserContext()
 
@@ -68,15 +69,16 @@ func (handler *ProductHandler) CreateProduct(i any, c *fiber.Ctx) error {
 
 // GetAllProducts is a method that retrieves all Products.
 //
-//		@Summary        Get all Products
-//		@Description    Get all Products with the input metadata or without metadata
-//		@Tags           Products
-//		@Produce        json
-//	 @Param 			organization_id path string true "Organization ID"
-//	 @Param 			ledger_id path string true "Ledger ID"
-//		@Param          metadata query string false "Metadata"
-//		@Success        200 {object} mpostgres.Pagination{items=[]mmodel.Product}
-//		@Router         /v1/organizations/{organization_id}/ledgers/{ledger_id}/products [get]
+//	@Summary		Get all Products
+//	@Description	Get all Products with the input metadata or without metadata
+//	@Tags			Products
+//	@Produce		json
+//	@Param			organization_id	path		string	true	"Organization ID"
+//	@Param			ledger_id		path		string	true	"Ledger ID"
+//	@Param			metadata		query		string	false	"Metadata"
+//	@Param			Midaz-Id		header		string							false	"Request ID"
+//	@Success		200				{object}	mpostgres.Pagination{items=[]mmodel.Product}
+//	@Router			/v1/organizations/{organization_id}/ledgers/{ledger_id}/products [get]
 func (handler *ProductHandler) GetAllProducts(c *fiber.Ctx) error {
 	ctx := c.UserContext()
 
@@ -138,15 +140,16 @@ func (handler *ProductHandler) GetAllProducts(c *fiber.Ctx) error {
 
 // GetProductByID is a method that retrieves Product information by a given id.
 //
-//		@Summary        Get a Product by ID
-//		@Description    Get a Product with the input ID
-//		@Tags           Products
-//		@Produce        json
-//	 @Param 			organization_id path string true "Organization ID"
-//	 @Param 			ledger_id path string true "Ledger ID"
-//		@Param          id path string true "Product ID"
-//		@Success        200 {object} mmodel.Product
-//		@Router         /v1/organizations/{organization_id}/ledgers/{ledger_id}/products/{id} [get]
+//	@Summary		Get a Product by ID
+//	@Description	Get a Product with the input ID
+//	@Tags			Products
+//	@Produce		json
+//	@Param			organization_id	path		string	true	"Organization ID"
+//	@Param			ledger_id		path		string	true	"Ledger ID"
+//	@Param			id				path		string	true	"Product ID"
+//	@Param			Midaz-Id		header		string							false	"Request ID"
+//	@Success		200				{object}	mmodel.Product
+//	@Router			/v1/organizations/{organization_id}/ledgers/{ledger_id}/products/{id} [get]
 func (handler *ProductHandler) GetProductByID(c *fiber.Ctx) error {
 	ctx := c.UserContext()
 
@@ -177,17 +180,18 @@ func (handler *ProductHandler) GetProductByID(c *fiber.Ctx) error {
 
 // UpdateProduct is a method that updates Product information.
 //
-//			@Summary        Update a Product
-//			@Description    Update a Product with the input payload
-//			@Tags           Products
-//			@Accept         json
-//			@Produce        json
-//	 	@Param 			organization_id path string true "Organization ID"
-//	 	@Param 			ledger_id path string true "Ledger ID"
-//		 	@Param 			id path string true "Product ID"
-//			@Param          product body mmodel.UpdateProductInput true "Product"
-//			@Success        200 {object} mmodel.Product
-//			@Router         /v1/organizations/{organization_id}/ledgers/{ledger_id}/products/{id} [patch]
+//	@Summary		Update a Product
+//	@Description	Update a Product with the input payload
+//	@Tags			Products
+//	@Accept			json
+//	@Produce		json
+//	@Param			organization_id	path		string						true	"Organization ID"
+//	@Param			ledger_id		path		string						true	"Ledger ID"
+//	@Param			id				path		string						true	"Product ID"
+//	@Param			product			body		mmodel.UpdateProductInput	true	"Product"
+//	@Param			Midaz-Id		header		string							false	"Request ID"
+//	@Success		200				{object}	mmodel.Product
+//	@Router			/v1/organizations/{organization_id}/ledgers/{ledger_id}/products/{id} [patch]
 func (handler *ProductHandler) UpdateProduct(i any, c *fiber.Ctx) error {
 	ctx := c.UserContext()
 
@@ -237,14 +241,15 @@ func (handler *ProductHandler) UpdateProduct(i any, c *fiber.Ctx) error {
 
 // DeleteProductByID is a method that removes Product information by a given ids.
 //
-//		@Summary        Delete a Product by ID
-//		@Description    Delete a Product with the input ID
-//		@Tags           Products
-//	 @Param 			organization_id path string true "Organization ID"
-//	 @Param 			ledger_id path string true "Ledger ID"
-//		@Param          id path string true "Product ID"
-//		@Success        204
-//		@Router         /v1/organizations/{organization_id}/ledgers/{ledger_id}/products/{id} [delete]
+//	@Summary		Delete a Product by ID
+//	@Description	Delete a Product with the input ID
+//	@Tags			Products
+//	@Param			organization_id	path	string	true	"Organization ID"
+//	@Param			ledger_id		path	string	true	"Ledger ID"
+//	@Param			id				path	string	true	"Product ID"
+//	@Param			Midaz-Id		header		string							false	"Request ID"
+//	@Success		204
+//	@Router			/v1/organizations/{organization_id}/ledgers/{ledger_id}/products/{id} [delete]
 func (handler *ProductHandler) DeleteProductByID(c *fiber.Ctx) error {
 	ctx := c.UserContext()
