+++ components/ledger/internal/bootstrap/http/routes.go
@@ -84,8 +84,7 @@ func NewRouter(lg mlog.Logger, tl *mopentelemetry.Telemetry, cc *mcasdoor.Casdoo
 	f.Get("/version", http.Version)
 
 	// Doc
-	f.Get("/swagger/*", http.WithSwaggerEnvConfig(), fiberSwagger.WrapHandler)
-	http.DocAPI("ledger", "Ledger API", f)
+	f.Get("/swagger/*", WithSwaggerEnvConfig(), fiberSwagger.WrapHandler)
 
 	f.Use(tlMid.EndTracingSpans)
