+++ components/ledger/internal/bootstrap/http/swagger.go
@@ -1,0 +1,34 @@
+package http
+
+import (
+	"github.com/LerianStudio/midaz/components/ledger/api"
+	"github.com/gofiber/fiber/v2"
+	"os"
+)
+
+// WithSwaggerEnvConfig sets the Swagger configuration for the API documentation from environment variables if they are set.
+func WithSwaggerEnvConfig() fiber.Handler {
+	return func(c *fiber.Ctx) error {
+		envVars := map[string]*string{
+			"SWAGGER_TITLE":       &api.SwaggerInfo.Title,
+			"SWAGGER_DESCRIPTION": &api.SwaggerInfo.Description,
+			"SWAGGER_VERSION":     &api.SwaggerInfo.Version,
+			"SWAGGER_HOST":        &api.SwaggerInfo.Host,
+			"SWAGGER_BASE_PATH":   &api.SwaggerInfo.BasePath,
+			"SWAGGER_LEFT_DELIM":  &api.SwaggerInfo.LeftDelim,
+			"SWAGGER_RIGHT_DELIM": &api.SwaggerInfo.RightDelim,
+		}
+
+		for env, field := range envVars {
+			if value := os.Getenv(env); value != "" {
+				*field = value
+			}
+		}
+
+		if schemes := os.Getenv("SWAGGER_SCHEMES"); schemes != "" {
+			api.SwaggerInfo.Schemes = []string{schemes}
+		}
+
+		return c.Next()
+	}
+}
